third_party/blink/web_tests/fast/css/invalidation/removed-hover-shadow-rule-expected.txt

Removing a shadow tree stylesheet should cause an invalidation set update.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.internals is defined.
PASS window.eventSender is defined.
PASS getComputedStyle(match).backgroundColor is red
PASS internals.updateStyleAndReturnAffectedElementCount() is 2
PASS getComputedStyle(match).backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE

Hover this text ... and this should have a green background.
